+++ tools/tpm2_nvincrement.c
@@ -13,33 +13,29 @@ struct tpm_nvincrement_ctx {
         tpm2_loaded_object object;
     } auth_hierarchy;
 
+    bool is_auth_hierarchy_specified;
     TPM2_HANDLE nv_index;
 };
-static tpm_nvincrement_ctx ctx;
+static tpm_nvincrement_ctx ctx = {
+    .is_auth_hierarchy_specified = false,
+};
+
+static bool on_arg(int argc, char **argv) {
+    /* If the users doesn't specify an authorization hierarchy use the index
+    * passed to -x/--index for the authorization index.
+    */
+    if (!ctx.is_auth_hierarchy_specified) {
+        ctx.auth_hierarchy.ctx_path = argv[0];
+    }
+    return on_arg_nv_index(argc, argv, &ctx.nv_index);
+}
 
 static bool on_option(char key, char *value) {
-    bool result;
 
     switch (key) {
-    case 'x':
-        result = tpm2_util_string_to_uint32(value, &ctx.nv_index);
-        if (!result) {
-            LOG_ERR("Could not convert NV index to number, got: \"%s\"",
-                    value);
-            return false;
-        }
-
-        if (ctx.nv_index == 0) {
-            LOG_ERR("NV Index cannot be 0");
-            return false;
-        }
-        /* If the users doesn't specify an authorisation hierarchy use the index
-        * passed to -x/--index for the authorisation index.
-        */
-        ctx.auth_hierarchy.ctx_path = value;
-        break;
     case 'C':
         ctx.auth_hierarchy.ctx_path = value;
+        ctx.is_auth_hierarchy_specified = true;
         break;
     case 'P':
         ctx.auth_hierarchy.auth_str = value;
@@ -53,13 +49,12 @@ static bool on_option(char key, char *value) {
 bool tpm2_tool_onstart(tpm2_options **opts) {
 
     const struct option topts[] = {
-        { "index",                required_argument, NULL, 'x' },
         { "hierarchy",            required_argument, NULL, 'C' },
         { "auth",                 required_argument, NULL, 'P' },
     };
 
-    *opts = tpm2_options_new("x:C:P:", ARRAY_LEN(topts), topts,
-                             on_option, NULL, 0);
+    *opts = tpm2_options_new("C:P:", ARRAY_LEN(topts), topts,
+                             on_option, on_arg, 0);
 
     return *opts != NULL;
 }